# MSR.Answer.API/spec/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "MSR API",
    "description": "REST API Answer 3",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://dev-api.answer.msr-fsr.com"
    }
  ],
  "paths": {
    "/v{version}/Account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_Login",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemLoginRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditActionResultOfString"
                }
              }
            }
          }
        }
      }
    },
    "/v{version}/Account/forgotpassword": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_ForgotPassword",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/v{version}/Account/forgotusername": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_ForgotUserName",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotUserNameRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v{version}/Account/resetpassword": {
      "patch": {
        "tags": [
          "Account"
        ],
        "operationId": "Account_ResetPassword",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditActionResultOfString": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuditActionResult"
          },
          {
            "type": "object",
            "additionalProperties": true,
            "properties": {
              "object": {
                "type": "string",
                "nullable": true
              },
              "returnedObject": {
                "nullable": true
              }
            }
          }
        ]
      },
      "AuditActionResult": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "successMessage": {
            "type": "string",
            "nullable": true
          },
          "errorMessages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ErrorMessage"
            }
          },
          "returnedObject": {
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "hasErrors": {
            "type": "boolean"
          },
          "hasValidationErrors": {
            "type": "boolean"
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "isValidationMessage": {
            "type": "boolean"
          }
        }
      },
      "SystemLoginRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "userName",
          "password"
        ],
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
            "userName": "dummy",
            "password": "dummy"
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "userName"
        ],
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
            "userName": "91b3e0455f51f25b"
        }
      },
      "ForgotUserNameRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
            "email": "91b3e0455f51f25b@cmhworks.com"
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "token",
          "newPassword"
        ],
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          },
          "newPassword": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
            "token": "kvcNhL+3YY5K4WfYNxC31POQcgSvdpsnSgrsfqeAncE=",
            "newPassword": "dummy"
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Copy this into the value field: Bearer {token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}